Finished splitting off the functions
