+++ Functions/Misc.py
@@ -1,6 +1,12 @@
 from csv import writer, reader
 from os.path import isfile
 
+from Functions.NewSuper import createSuper
+from Functions.FindTag import findTag
+from Functions.SelectSection import selectSection
+from Functions.Tesys import createTesys
+from Functions.Misc import findFile, createFile, checkAnother
+
 
 def findFile():
     """Finds the base CSV file and loops if not correct"""
@@ -42,3 +48,72 @@ def createFile(override):
             return ""
         else:
             print("Error: Expected answer \"yes\" or \"no\"")
+
+
+def checkAnother(item):
+    """Checks if more than one input is required"""
+    moreInputs = input("Another "+item+"? (Y/N) ").lower().replace(" ", "")
+    if moreInputs in ["y", "ye", "yes", "1", "true"]:
+        return True
+    elif moreInputs in ["n", "no", "0", "false"]:
+        return False
+    else:
+        print("Error: \"Yes\" or \"No\" answer required")
+        return checkAnother(item)
+
+
+def findFunction(functionCheck):
+    """Finds the function required, and gathers/passes the information needed to run"""
+    # Function that finds a given tag in a database, and optionally saves it to a file
+    if functionCheck in ["findtag", "ft"]:
+        fileName = findFile()
+        moreTags = True
+        tagList = []
+        while moreTags == True:
+            requiredTag = input("Tag to search for: ").lower()
+            if requiredTag.startswith(":"):
+                requiredTag = requiredTag[1:]
+            else:
+                tagList += [requiredTag]
+            moreTags = checkAnother("tag")
+
+        print(findTag(fileName, createFile(True), tagList))
+
+    # Function that selects a certain section of tags, and saves them only to a file
+    elif functionCheck in ["selectsection", "selsec", "ss"]:
+        fileName = findFile()
+        # Discovers which section is required
+        moreSections = True
+        sections = []
+        while moreSections == True:
+            sections.append(input("Which section to keep: ").lower())
+            moreSections = checkAnother("section")
+
+        print(selectSection(fileName, createFile(True), sections))
+
+    elif functionCheck in ["super"]:
+
+        print(createSuper())
+
+    elif functionCheck in ["tesys"]:
+        tesysLoop = True
+        tesysList = []
+        tesysNum = 1
+        while tesysLoop == True:
+            tesysListTemp = {}
+            tesysListTemp["Tag"] = input("Tag #"+str(tesysNum)+": ")
+            tesysListTemp["Comment"] = input("Comment #"+str(tesysNum)+": ")
+            tesysListTemp["AccessName"] = input(
+                "AccessName #"+str(tesysNum)+": ")
+            tesysList.append(tesysListTemp)
+            tesysNum += 1
+            tesysLoop = checkAnother("set of tags")
+
+        print(createTesys(createFile(True), tesysList))
+
+    # Loops if the given function doesn't exist/isn't recognised
+    else:
+        print("Functions: \"Find Tag\", \"Select Section\", \"Super\", \"Tesys\"")
+        findFunction(
+            input("Function type required: ").lower().replace(" ", ""))
+    return
